sso/files/requirements.txt
astroid==1.2.1
certifi==14.5.14
Django==1.7.8
django-braces==1.4.0
django-cors-headers==0.13
# django-json-field==0.5.7
-e git+https://github.com/cro/django-pgjsonb.git#egg=django-pgjsonb
django-mptt==0.6.1
djangorestframework==2.4.2
gunicorn==19.1.1
httplib2==0.8
Jinja2==2.7.3
logilab-common==0.63.0
MarkupSafe==0.23
msgpack-python==0.4.5
pycrypto==2.6.1
# PyMySQL==0.6.2
psycopg2==2.6
python-dateutil==2.2
pytz==2014.7
PyYAML==3.11
requests==2.2.1
-e git+https://github.com/saltstack/salt.git@d2ec0e0f8d90f78a478e5dca81fcab1a1c65e9cf#egg=salt-develop
six==1.8.0
sqlparse==0.1.14
tornado==4.0.2
netifaces==0.10.4
django-extensions==1.5.4
